+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - disable autocompletion for hex digits
 - no more crashes on recovered parse result
 - correct handling of using '=' sign in expressions
+- work with untitled (unsaved) documents
 
 ## [1.3.1] - 2023-07-11